Refactored Gecko node to use TestSuiteRunner
